Build out of Epic search, create position, load position and transactions

+++ Session.py
@@ -12,10 +12,13 @@ class SessionManager:
 
     def __init__(self):
         self.session_headers = None
+        self.session_account = None
         load_dotenv()
         self.username = os.environ.get("IG_USERNAME")
         self.password = os.environ.get("IG_PASSWORD")
         self.apiKey = os.environ.get("IG_API_KEY")
+        self.CFD_account = os.environ.get("CFDACC")
+        self.spread_account = os.environ.get("SPREADACC")
         self.live_session = False
         print("SESSION MANAGER INITIALISED: ENV variables= ", self.username, self.password, self.apiKey)
 
@@ -27,6 +30,11 @@ class SessionManager:
     def dump_session_json(self):
         with open('sessiontokens.json', 'w') as f:
             json.dump(self.session_headers, f)
+
+    def what_account(self):
+        MAP = {"Z4ZGAF": "CFD", "Z4ZGAG" : "SPREAD BET"}
+        print("ACCOUNT USED ON THIS SESSION IS: ", self.session_account, MAP)
+        return
 
     def create_live_session(self):
         if self.live_session == True:
@@ -82,6 +90,7 @@ class SessionManager:
             self.session_headers["CST"] = r.headers["CST"]
             self.dump_session_json()
             self.live_session = True
+            self.session_account = json.loads(r.content.decode("utf-8"))['accountType']
             print("SUCCESSFUL LOGIN AND DUMP OF SESSION HEADERS: ", r.content.decode("utf-8"))
 
         else:
@@ -94,10 +103,41 @@ class SessionManager:
         r = requests.request(method=method, url=url, headers=self.session_headers, json=body)
         return r
 
+    def change_account(self, account="CFD", to_default=True):
+        """ Use 'CFD' or 'SPREAD' as string arg to account"""
+
+        if account == "CFD":
+            openacc = self.CFD_account
+        else:
+            openacc = self.spread_account
+
+        if not self.live_session:
+            self.create_live_session()
+
+        url = "https://demo-api.ig.com/gateway/deal/session"
+        version = "1"
+        body = {"accountId": openacc,
+                "defaultAccount": to_default
+                }
+
+        r = self.API_call('PUT', url=url, body=body, version=version)
+
+        if r.status_code == 200:
+            print("SUCCESSFUL ACCOUNT SWITCH ", r.status_code, r.content.decode("utf-8"), r.headers)
+            self.session_headers["X-SECURITY-TOKEN"] = r.headers["X-SECURITY-TOKEN"]
+            self.dump_session_json()
+            self.session_account = openacc
+        else:
+            print("ERROR ON ACCOUNT SWITCH: ", r.status_code, r.content.decode("utf-8"))
+        return
+
     def logout(self):
         url = "https://demo-api.ig.com/gateway/deal/session"
-        self.create_live_session()
         version = "1"
+
+        if not self.live_session:
+            print("NO LIVE SESSION: WILL NOT CALL API TO LOG OUT")
+            return
 
         r = self.API_call(method="DELETE", url=url, body={}, version=version)
